+++ app/assets/javascripts/Results/keybinding.js
@@ -22,11 +22,11 @@ Mousetrap.bind("shift+right", function () {
   }
 });
 
-// Go to random submission with ctrl + shift + right
+// Go to random incomplete submission with ctrl + shift + right
 Mousetrap.bind("ctrl+shift+right", function () {
   // Don't override range selection keybindings
   if (!is_text_selected()) {
-    $(".button.random-unmarked-submission")[0].click();
+    $(".button.random-incomplete-submission")[0].click();
   }
 });